refactor(lambda): split extras.rs by surface (2851 -> 1198 mod + 10 submodules)
Audit 2026-05-19 P9 continuation. Lambda's extras.rs (which collects
the 62 non-core ops beyond function CRUD/invoke) split into surface-
keyed submodules under extras/:

- layers.rs          404  (publish/list/get/delete + permissions)
- concurrency.rs     237  (reserved + provisioned + scaling-config)
- stream.rs          195  (invoke_with_response_stream)
- code_signing.rs    193  (config CRUD + function code-signing assignment)
- function_url.rs    171  (URL config CRUD)
- account.rs         155  (account settings + handle_extra dispatcher)
- aliases.rs         154
- event_invoke.rs    115  (event_invoke config)
- runtime.rs          94  (runtime_management)
- recursion.rs        57  (recursion_config)

151 unit tests pass.
